# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API sobre el Sistema de Transporte Colectivo (STC) de la Ciudad de México",
        "title": "Apimetro",
        "contact": {
            "name": "galigaribaldi (Galileo Cabrera Garibaldi)",
            "email": "galigaribaldi0@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/movilidad",
    "paths": {
...
        "/{sistema}/descripcion-estacion": {
            "get": {
                "description": "Obtener descripciones de estaciones con filtros",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DescripcionEstacion"
                ],
                "summary": "Datos de Descripcion Estacion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sistema de transporte",
                        "name": "sistema",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID de la descripción",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by nombre",
                        "name": "nombre",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by alcaldia_municipio",
                        "name": "alcaldia_municipio",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by num_comercial",
                        "name": "num_comercial",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.DescripcionEstacion"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Crear una nueva descripción de estación",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DescripcionEstacion"
                ],
                "summary": "Crear Descripcion Estacion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sistema de transporte",
                        "name": "sistema",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos de la descripción de estación",
                        "name": "descripcion",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DescripcionEstacion"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        },
        "/{sistema}/descripcion-estacion/{id}": {
            "put": {
                "description": "Actualizar una descripción de estación por ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DescripcionEstacion"
                ],
                "summary": "Actualizar Descripcion Estacion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sistema de transporte",
                        "name": "sistema",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID de la descripción de estación",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos actualizados",
                        "name": "descripcion",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DescripcionEstacion"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Eliminar una descripción de estación por ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DescripcionEstacion"
                ],
                "summary": "Eliminar Descripcion Estacion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sistema de transporte",
                        "name": "sistema",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID de la descripción de estación a eliminar",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "httputil.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "models.DescripcionEstacion": {
            "type": "object",
            "properties": {
                "alcaldia_municipio": {
                    "type": "string"
                },
                "anio": {
                    "type": "string"
                },
                "cve_est": {
                    "type": "string"
                },
                "descripcion_estacion_id": {
                    "type": "integer"
                },
                "estacion_id": {
                    "type": "integer"
                },
                "estado_ciudad": {
                    "type": "string"
                },
                "longitud": {
                    "type": "number"
                },
                "nombre": {
                    "type": "string"
                },
                "tipo": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}